requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
curl-cffi>=0.7.0
seleniumbase>=4.0.0
